backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
pydantic-settings==2.2.1
structlog==24.1.0
slowapi==0.1.9
google-cloud-aiplatform
supabase==2.4.1
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
